requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.2.0
cryptography==42.0.4
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
ecdsa==0.19.0
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.110.0
fastapi-sessions==0.3.2
fastjsonschema==2.19.1
fonttools==4.45.0
fqdn==1.5.1
frozenlist==1.4.1
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.12.5
jupyter_server_proxy==4.1.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/macholib-1.15.2-py2.py3-none-any.whl
mariadb==1.1.8
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
memory-profiler==0.61.0
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.2
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prometheus_client==0.20.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-jose==3.3.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scipy==1.12.0
Send2Trash==1.8.2
simpervisor==1.0.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
soupsieve==2.5
spicy==0.16.0
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
traitlets==5.11.2
txaio==23.1.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.2.1
utils==1.0.1
uvicorn==0.27.1
vp==0.0.1
vpython==7.6.4
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.10
yarl==1.9.4
zipp==3.17.0
